requirements.txt
-i https://pypi.org/simple
allure-pytest==2.15.3
allure-python-commons==2.15.3; python_version >= '3.6'
attrs==25.4.0; python_version >= '3.9'
certifi==2026.1.4; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
greenlet==3.3.0; python_version >= '3.10'
idna==3.11; python_version >= '3.8'
iniconfig==2.3.0; python_version >= '3.10'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.3; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
playwright==1.57.0; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
pyee==13.0.0; python_version >= '3.8'
pygments==2.19.2; python_version >= '3.8'
pytest==9.0.2; python_version >= '3.10'
pytest-base-url==2.1.0; python_version >= '3.8'
pytest-html==4.1.1; python_version >= '3.8'
pytest-metadata==3.1.1; python_version >= '3.8'
pytest-playwright==0.7.2; python_version >= '3.10'
python-dotenv==1.2.1; python_version >= '3.9'
python-slugify==8.0.4; python_version >= '3.7'
requests==2.32.5; python_version >= '3.9'
text-unidecode==1.3
typing-extensions==4.15.0; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
